Apply patch to silence some compiler warnings.
Submitted by: Patrick Hartling


git-svn-id: https://opensg.vrsource.org/svn/trunk@696 4683daeb-ad0f-0410-a623-93161e962ae5

+++ Source/System/RenderingBackend/OSGStageValidator.inl
@@ -51,7 +51,7 @@ StageValidator::ValidationStatus StageValidator::validate(Int32 iStageId)
     if(iStageId < 0)
         return Self::Unknown;
 
-    if(_vStatusStore.size() <= iStageId)
+    if(_vStatusStore.size() <= static_cast<UInt32>(iStageId))
     {
         StageStatus tmpStat;
